+++ src-tauri/src/profile/import/mod.rs
@@ -15,11 +15,11 @@ use reqwest::StatusCode;
 use serde::{Deserialize, Serialize};
 use tauri::AppHandle;
 use tempfile::tempdir;
-use tracing::{debug, info, trace, warn};
+use tracing::{info, trace, warn};
 use uuid::Uuid;
-use walkdir::WalkDir;
 
 use crate::{
+    game::Game,
     prefs::Backends,
     profile::{
         export::{PROFILE_DATA_PREFIX, ProfileManifest},
@@ -156,7 +156,8 @@ pub(super) async fn import_profile(
         "importing profile"
     );
 
-    let (profile_id, profile_path, to_install) = prepare_import(&options, data.manifest, app)?;
+    let (profile_id, profile_path, game, to_install) =
+        prepare_import(&options, data.manifest, app)?;
 
     let result = app
         .install_queue()
@@ -165,7 +166,13 @@ pub(super) async fn import_profile(
 
     let result = match result {
         Ok(()) => {
-            import_config(&profile_path, &data.path, &options).context("error importing config")?;
+            import_config(
+                &profile_path,
+                &data.path,
+                game.mod_loader.mod_config_dirs(),
+                &options,
+            )
+            .context("error importing config")?;
 
             Ok(profile_id)
         }
@@ -194,7 +201,7 @@ fn prepare_import(
     options: &ImportOptions,
     manifest: ProfileManifest,
     app: &AppHandle,
-) -> Result<(i64, PathBuf, Vec<ModInstall>)> {
+) -> Result<(i64, PathBuf, Game, Vec<ModInstall>)> {
     let ProfileManifest {
         name,
         mods,
@@ -236,7 +243,7 @@ fn prepare_import(
 
     game.save(app)?;
 
-    Ok((id, path, to_install))
+    Ok((id, path, game.game, to_install))
 }
 
 fn cleanup_failed_profile(profile_id: i64, app: &AppHandle) -> Result<()> {
@@ -315,22 +322,30 @@ fn incremental_update(
     Ok(to_install)
 }
 
-pub fn import_config(dest: &Path, src: &Path, options: &ImportOptions) -> Result<()> {
-    let src_files = WalkDir::new(src)
-        .follow_links(false)
-        .into_iter()
-        .filter_map(Result::ok)
-        .filter(|entry| entry.file_type().is_file())
-        .map(|entry| {
-            entry
-                .into_path()
-                .strip_prefix(src)
-                .expect("path should be child of source")
-                .to_path_buf()
-        })
-        .filter(|path| options.import_all || is_always_included(path));
+#[tracing::instrument(skip_all, fields(dest = %dest.display(), src = %src.display()))]
+pub fn import_config(
+    dest: &Path,
+    src: &Path,
+    config_dirs: &[&str],
+    options: &ImportOptions,
+) -> Result<()> {
+    let src_files: HashSet<PathBuf> = super::export::list_files(src)
+        .filter(|path| options.import_all || is_always_imported(path))
+        .collect();
 
-    debug!("importing config files from source to destination");
+    let dest_files: HashSet<PathBuf> = super::export::find_config(dest, config_dirs).collect();
+
+    if !options.merge {
+        // remove existing extra config files that are not in the imported profile
+        for extra_file in dest_files.difference(&src_files) {
+            let extra_path = dest.join(extra_file);
+            trace!(
+                relative_path = %extra_file.display(),
+                "removing extra config file"
+            );
+            fs::remove_file(extra_path).fs_context("removing extra config file", &extra_file)?;
+        }
+    }
 
     for file in src_files {
         let src_path = src.join(&file);
@@ -364,7 +379,7 @@ pub fn import_config(dest: &Path, src: &Path, options: &ImportOptions) -> Result
     Ok(())
 }
 
-fn is_always_included(path: impl AsRef<Path>) -> bool {
+fn is_always_imported(path: impl AsRef<Path>) -> bool {
     static EXCLUDE_SET: LazyLock<GlobSet> = LazyLock::new(|| {
         GlobSetBuilder::new()
             .add(Glob::new("export.r2x").unwrap())